+++ src/css/blockList.css
@@ -1,8 +1,17 @@
 .blockList {
-    background-color: red;
+    /* background-color: grey; */
     width: 100%;
     display: flex;
-
-    /* flex-direction: row, */
-    
+    justify-content: center;
 }
+
+@media only screen and (max-width: 798px) {
+    .blockList {
+        justify-content: center;
+        display: flex;
+        flex-flow: column;
+        
+    }   
+}
+
+
