+++ src/index.js
@@ -7,7 +7,6 @@ import { Provider } from "react-redux";
 import { createStore, applyMiddleware} from "redux";
 import thunk from "redux-thunk";
 import App from './App';
-import Card from "./components/Card"
 
 const store = createStore(reducer,applyMiddleware(thunk));
